+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [2022-05-30]
+### Upgraded
+- "Keyple Util Library" to version `2.1.0`
+
 ## [2021-12-20]
 ### Upgraded
 - "Keyple Card Generic Library" to version `2.0.2`
@@ -32,7 +36,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - org.eclipse.keyple:keyple-card-generic-java-lib:2.0.0
   - org.eclipse.keyple:keyple-util-java-lib:2.+
 
-[unreleased]: https://github.com/eclipse/keyple-integration-java-test/compare/2021-12-08...HEAD
+[unreleased]: https://github.com/eclipse/keyple-integration-java-test/compare/2022-05-30...HEAD
+[2022-05-30]: https://github.com/eclipse/keyple-integration-java-test/compare/2021-12-20...2022-05-30
+[2021-12-20]: https://github.com/eclipse/keyple-integration-java-test/compare/2021-12-08...2021-12-20
 [2021-12-08]: https://github.com/eclipse/keyple-integration-java-test/compare/2021-11-22...2021-12-08
 [2021-11-22]: https://github.com/eclipse/keyple-integration-java-test/compare/2021-10-06...2021-11-22
 [2021-10-06]: https://github.com/eclipse/keyple-integration-java-test/releases/tag/2021-10-06